add SoundManager
사운드 추가(배경음, 플레이어 효과음), 오프닝이랑 엔딩 컷씬 텍스트 타이핑 효과 추가
