--- Assets/Scripts/GameManager.cs
@@ -1,37 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class GameManager : MonoBehaviour
-{
-    private static GameManager instance;
-
-    public static GameManager Instance
-    {
-        get
-        {
-            return instance;
-        }
-    }
-
-    private void Awake()
-    {
-        if(instance)
-        {
-            Destroy(gameObject);
-            return;
-        }
-
-        instance = this;
-    }
-
-    void Start()
-    {
-        
-    }
-
-    void Update()
-    {
-        
-    }
-}
