+++ Assets/Sounds.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5242e115ba566a7408e85f25d1db5f72
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
